remove the old billing UI
