+++ frontend/src/scenes/billing/Billing.stories.tsx
@@ -1,10 +1,9 @@
 import { Meta } from '@storybook/react'
 import { Billing } from './Billing'
-import { useStorybookMocks, mswDecorator, useFeatureFlags } from '~/mocks/browser'
+import { useStorybookMocks, mswDecorator } from '~/mocks/browser'
 import preflightJson from '~/mocks/fixtures/_preflight.json'
 import billingJson from '~/mocks/fixtures/_billing_v2.json'
 import billingJsonWithDiscount from '~/mocks/fixtures/_billing_v2_with_discount.json'
-import { FEATURE_FLAGS } from 'lib/constants'
 
 export default {
     title: 'Scenes-Other/Billing v2',
@@ -28,7 +27,6 @@ export default {
 } as Meta
 
 export const _BillingV2 = (): JSX.Element => {
-    useFeatureFlags([FEATURE_FLAGS.BILLING_BY_PRODUCTS])
     useStorybookMocks({
         get: {
             '/api/billing-v2/': {
@@ -41,7 +39,6 @@ export const _BillingV2 = (): JSX.Element => {
 }
 
 export const BillingV2WithDiscount = (): JSX.Element => {
-    useFeatureFlags([FEATURE_FLAGS.BILLING_BY_PRODUCTS])
     useStorybookMocks({
         get: {
             '/api/billing-v2/': {
